# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateRecurrencePatternRequest:
    properties:
      end_date:
        type: string
      frequency:
        enum:
        - daily
        - weekly
        - monthly
        - yearly
        type: string
      interval:
        minimum: 1
        type: integer
      transaction_id:
        type: string
    required:
    - frequency
    - interval
    - transaction_id
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  dto.GenerateRecurrenceRequest:
    properties:
      until_date:
        description: Gerar transações até esta data
        type: string
    required:
    - until_date
    type: object
...
  dto.RecurrencePatternDTO:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      frequency:
        type: string
      id:
        type: string
      interval:
        type: integer
      last_generated_date:
        type: string
      transaction_id:
        type: string
      user_id:
        type: string
    type: object
...
  dto.UpdateRecurrencePatternRequest:
    properties:
      end_date:
        type: string
      frequency:
        enum:
        - daily
        - weekly
        - monthly
        - yearly
        type: string
      interval:
        minimum: 1
        type: integer
    required:
    - frequency
    - interval
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API do sistema Organiza Aqui
  title: Organiza Aqui API
  version: "1.0"
paths:
...
  /api/v1/recurrence:
    get:
      description: Retorna todos os padrões de recorrência do usuário
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RecurrencePatternDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Listar padrões de recorrência
      tags:
      - recurrence
    post:
      consumes:
      - application/json
      description: Cria um padrão de recorrência para uma transação
      parameters:
      - description: Dados do padrão
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRecurrencePatternRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RecurrencePatternDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar padrão de recorrência
      tags:
      - recurrence
  /api/v1/recurrence/{id}:
    delete:
      description: Remove um padrão de recorrência
      parameters:
      - description: ID do padrão
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletar padrão de recorrência
      tags:
      - recurrence
    get:
      description: Retorna um padrão de recorrência específico
      parameters:
      - description: ID do padrão
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RecurrencePatternDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter padrão de recorrência
      tags:
      - recurrence
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um padrão de recorrência
      parameters:
      - description: ID do padrão
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRecurrencePatternRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RecurrencePatternDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar padrão de recorrência
      tags:
      - recurrence
  /api/v1/recurrence/generate:
    post:
      consumes:
      - application/json
      description: Gera transações baseadas nos padrões de recorrência até uma data
      parameters:
      - description: Data limite
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GenerateRecurrenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Gerar transações recorrentes
      tags:
      - recurrence
...
swagger: "2.0"